change permission requirements on config file
Don't allow the group to have write access to the main config file.
Although this might be used safely, it is not standard Unix practice to
delegate permissions this way. Furthermore, now that all secrets have
moved to their own file, it's ok if the group or others have read access
to the main config file.