support queueing of multiple packets
Previously only one packet could be queued. If multiple packets
we're sent out the tunnel device while a valid session did not
exist, only the last packets would be queued. Now all packets are
queued using a singly linked list and sent as soon as a valid session
is established.