pre-allocate tentative and next sessions
Pre-allocate tentative and next sessions so that no new memory is
allocated as long as a session is not authenticated. The reason we
still want to allocate new memory once a new session is established
is because we want to put the new session secrets on a new random
place on the heap.