refactor local address selection
Simplify connecting to roaming peers and peers with known endpoints.
Unprivileged processes can not reconnect(2) to a roaming peer from an
ip-address that is an alias and that is also used by the server to
listen for incoming connections. The primary address of an interface
will always be used for outbound connections no matter to which socket a
peer sent its first packet.

The only way to work around this issue and allow interface aliases to be
used is by creating a new socket each time a peer roams so that the right
address can be bound to the new socket. This would require setting the
SO_REUSEPORT option, which can only be done if pledge(2) is not used.

We make the trade-off to not use any privileged processes and limit the
set of possible source addresses to interface primary addresses only,
when initiating a connection to a peer.