replace sockaddr_storage with a sockaddr_{in6,in} union in IPC
A sockaddr_storage structure can contain any kind of network address on
the local system, but takes 256 bytes. Replace this structure with a new
union between a sockaddr_in6 and a sockaddr_in structure in all places
that only need to hold an IP-address. This new union is only 28 bytes.
Since some IPC messages contain two addresses this greatly reduces the
number of bytes used in IPC messages.