properly support big-endian systems
Session ids are either stored in a uint32_t or an int64_t. When a
uint32_t is used, the session id is stored in wire format, which is
little-endian. When an int64_t is used the number is represented in host
byte order. 64 bit ids are used in places where -1 signifies that the
session id is not set, or when used as a timer id. Furthermore, all
messages that are sent or recieved over sockets or IPC channels are in
wire format.

* fixes initiator role on big-endian
* show the right port when parking a socket fails