lower tunnel MTU from 1420 to 1408
Previously, maximum sized data messages were incorrectly padded and
could yield 1504 byte packets which are dropped on most networks. By
simply lowering the tunnel device MTU to a more conservative 1408 the
maximum sized IPv6 UDP packet will be 1488 bytes. This has the added
benefit to better support DSL networks that often have an MTU of 1492
bytes because of PPPoE.

Future work should include better determination of the path MTU and
padded data should be capped to be at most the tunnel MTU but we're to
close to a release now to incorporate that change in this commit.